requirements.txt
apiclient==1.0.4
asgiref==3.3.4
bcrypt==3.2.0
bigquery==0.0.12
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dacktool==0.0.7
dbstream==0.0.30
discovery==0.0.4
Django==3.2
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.4
environ==1.0
google-api-core==1.26.3
google-api-python-client==1.7.11
google-auth==1.28.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googleauthentication==0.0.15
grpcio==1.36.1
httplib2==0.19.1
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
paramiko==2.7.2
proto-plus==1.18.1
protobuf==3.15.7
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.5.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
sshtunnel==0.1.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
